+++ src/pub/proxy_edit/fields.js
@@ -49,14 +49,7 @@ export const tabs = {
                 label: 'Operating System',
                 tooltip: 'Operating System',
             },
-            timezone: {
-                label: 'Timezone',
-                tooltip: 'Select a timezone to be used by the browser',
-            },
-            resolution: {
-                label: 'Resolution',
-                tooltip: 'Window resolution to be used by the browser',
-            },
+
         },
     },
     rules: {
@@ -208,15 +201,26 @@ export const tabs = {
             },
         },
     },
-    headers: {
-        label: 'Headers',
-        tooltip: `Set default headers that are sent with each request from this
-            proxy port`,
+    browser: {
+        label: 'Browser',
+        tooltip: 'Configure settings for the browser launched by LPM',
         fields: {
             ua: {
                 label: 'Unblocker Mobile UA',
                 tooltip: `Let Unblocker set a random User-Agent from a mobile
                     browser automatically`,
+            },
+            timezone: {
+                label: 'Timezone',
+                tooltip: 'Select a timezone to be used by the browser',
+            },
+            resolution: {
+                label: 'Resolution',
+                tooltip: 'Window resolution to be used by the browser',
+            },
+            webrtc: {
+                label: 'WebRTC',
+                tooltip: 'Choose how WebRTC plugin behaves in the browser',
             },
         },
     },