+++ test/manager.js
@@ -613,6 +613,8 @@ describe('manager', ()=>{
                 'America/Sao_Paulo');
             t('with custom resolution', {resolution: '800x600'},
                 'resolution', {width: 800, height: 600});
+            t('webrtc is disabled', {webrtc: 'disabled'}, 'webrtc',
+                'disabled');
         });
     });
     describe('flags', ()=>{